fix comparision

+++ api/oc_ri.c
@@ -1109,7 +1109,7 @@ oc_ri_invoke_coap_entity_handler(void *request, void *response, uint8_t *buffer,
         len_resource = oc_string_len(resource->uri);
         // incoming URL should be equal or larger than the one with the wild card 
         // comparison should match to what ever is in front of the last char.
-        if ( ((uri_path_len + 1) >= len_resource)  &&
+        if ( ((int)(uri_path_len + 1) >= len_resource)  &&
             strncmp((const char *)oc_string(resource->uri) + 1, uri_path,
                     len_resource - 2 ) == 0) {
           request_obj.resource = cur_resource = resource;
